backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
httpx==0.26.0
python-multipart==0.0.9
PyMuPDF==1.23.26
beautifulsoup4==4.12.3
requests==2.31.0
PyJWT==2.6.0
python-Levenshtein==0.25.0
fuzzywuzzy==0.18.0
langchain==0.1.9
sentence-transformers==2.5.1
numpy==1.26.4
pydantic==2.6.1
pdfminer.six==20231228
lxml==5.1.0
python-dotenv==1.0.1
aiohttp==3.9.3
xmltodict==0.13.0
passlib[bcrypt]==1.7.4
